Extract resource parameters commons
